fixed the query but need to get it so the page will reload upon login so the page can look like normal Account page not login screen
